Version Packages (canary) (#107)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.

⚠️⚠️⚠️⚠️⚠️⚠️

`main` is currently in **pre mode** so this branch has prereleases
rather than normal releases. If you want to exit prereleases, run
`changeset pre exit` on `main`.

⚠️⚠️⚠️⚠️⚠️⚠️

# Releases
## @whop/api@0.0.21-canary.7

### Patch Changes

-   6330be2: release initial checkout package

## @whop/checkout@0.0.21-canary.7

### Patch Changes

-   6330be2: release initial checkout package

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>