Version Packages (canary) (#116)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.

⚠️⚠️⚠️⚠️⚠️⚠️

`main` is currently in **pre mode** so this branch has prereleases
rather than normal releases. If you want to exit prereleases, run
`changeset pre exit` on `main`.

⚠️⚠️⚠️⚠️⚠️⚠️

# Releases
## @whop/api@0.0.21-canary.8

### Patch Changes

-   cb2da90: fix(deps): update dependency js-md5 to v0.8.3
-   116374a: chore(deps): update dependency yaml to v2.8.0
- b406605: remove purchaseUrl from chargerUser because it should not be
used

## @whop/checkout@0.0.21-canary.8

### Patch Changes

-   3055447: support data-whop-checkout-session-id for embedded checkout
-   116374a: chore(deps): update dependency yaml to v2.8.0

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>