Version Packages (#144)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/iframe@0.0.1

### Patch Changes

-   ad73b88: init `@whop/iframe` and `@whop/react` packages

## @whop/react@0.0.1

### Patch Changes

-   ad73b88: init `@whop/iframe` and `@whop/react` packages
-   Updated dependencies [ad73b88]
    -   @whop/iframe@0.0.1

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>