Version Packages (#157)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.23

### Patch Changes

-   9d57bfe: chore(deps): update dependency ts-proto to v2.7.1
-   a2809a2: chore(deps): update dependency turbo to v2.5.4
- 964850f: Added waitlist entries to the SDK along with if the user is
phone verified and their location (if permitted).
-   f8627a6: chore(deps): update dependency mintlify to v4.0.563
- ad76e87: Split the functions available on the client and server SDK
such that you cannot call mutations and queries on the client that will
always fail due to auth issues.

## @whop/checkout@0.0.23

### Patch Changes

-   9d57bfe: chore(deps): update dependency ts-proto to v2.7.1
-   a2809a2: chore(deps): update dependency turbo to v2.5.4
-   f8627a6: chore(deps): update dependency mintlify to v4.0.563

## @whop/react@0.0.3

### Patch Changes

-   d18861f: chore(deps): pin dependency @types/react to 19.1.5

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>