Version Packages (#165)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.24

### Patch Changes

-   1cd7d96: update .npmignore to include build configs

## @whop/checkout@0.0.24

### Patch Changes

-   1cd7d96: support `center` event and handle event listener cleanup
-   1cd7d96: update .npmignore to include build configs

## @whop/iframe@0.0.3

### Patch Changes

-   1cd7d96: update .npmignore to include build configs

## @whop/react@0.0.4

### Patch Changes

-   1cd7d96: update .npmignore to include build configs
-   Updated dependencies [1cd7d96]
    -   @whop/iframe@0.0.3

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>