Version Packages (#187)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.26

### Patch Changes

-   f23cc24: add search query to the list users for experience endpoint
-   bc3353e: add `oauth2` module to simplify oauth implementations

## @whop/react@0.2.1

### Patch Changes

-   b20a83e: add `WhopApp` wrapper component
-   f3e149b: add `withWhopAppConfig` wrapper for next

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>