Version Packages (#197)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.27

### Patch Changes

-   06c85ff: Add `hidePrice` prop to @whop/checkout and @whop/react

## @whop/react@0.2.3

### Patch Changes

-   06c85ff: Add `hidePrice` prop to @whop/checkout and @whop/react
-   Updated dependencies [06c85ff]
    -   @whop/checkout@0.0.27

## @local/docs@1.0.4



## @local/ai-image-generator@0.1.2



## @local/graphql-codegen-docs@0.0.5

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>