Version Packages (canary) (#201)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.

⚠️⚠️⚠️⚠️⚠️⚠️

`main` is currently in **pre mode** so this branch has prereleases
rather than normal releases. If you want to exit prereleases, run
`changeset pre exit` on `main`.

⚠️⚠️⚠️⚠️⚠️⚠️

# Releases
## @whop/react@0.2.4-canary.0

### Patch Changes

- 6836164: update `exports` field to allow for easier sub-package
discovery in editors

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>