Version Packages (#202)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/react@0.2.4

### Patch Changes

- 6836164: update `exports` field to allow for easier sub-package
discovery in editors

## @local/docs@1.0.5



## @local/ai-image-generator@0.1.3



## @local/graphql-codegen-docs@0.0.6

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>