Version Packages (#207)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.28

### Patch Changes

-   2e21c45: add `skipRedirect` option to checkout and react packages

## @whop/react@0.2.5

### Patch Changes

-   2e21c45: add `skipRedirect` option to checkout and react packages
-   Updated dependencies [2e21c45]
    -   @whop/checkout@0.0.28

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>