Version Packages (#213)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.29

### Patch Changes

-   c74b64a: add `onComplete` prop to `@whop/checkout` and `@whop/react`

## @whop/react@0.2.7

### Patch Changes

-   c74b64a: add `onComplete` prop to `@whop/checkout` and `@whop/react`
-   Updated dependencies [c74b64a]
    -   @whop/checkout@0.0.29

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>