Version Packages (#215)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.28

### Patch Changes

-   afc4d5a: Added the ability to create app builds via the api

## @whop/react@0.2.8

### Patch Changes

-   Updated dependencies [afc4d5a]
    -   @whop/api@0.0.28

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>