Version Packages (#218)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.29

### Patch Changes

-   551144b: add `Decimal` to gql scalar defs

## @whop/checkout@0.0.30

### Patch Changes

-   f3ed30f: add support for utm param forwarding

## @whop/react@0.2.9

### Patch Changes

-   f3ed30f: add support for utm param forwarding
-   Updated dependencies [f3ed30f]
-   Updated dependencies [551144b]
    -   @whop/checkout@0.0.30
    -   @whop/api@0.0.29

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>