Version Packages (#229)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/react@0.2.13

### Patch Changes

-   428bf11: reorder package.json exports

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>