Version Packages (#233)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.30

### Patch Changes

-   e41599c: Added list access passes for experience endpoint

## @whop/react@0.2.15

### Patch Changes

-   Updated dependencies [e41599c]
    -   @whop/api@0.0.30

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>