Version Packages (#235)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.31

### Patch Changes

-   7650fda: Added attachments to access pass

## @whop/react@0.2.16

### Patch Changes

-   Updated dependencies [7650fda]
    -   @whop/api@0.0.31

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>