Version Packages (#237)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.31

### Patch Changes

-   ba90230: add styling api to embedded checkout

## @whop/react@0.2.17

### Patch Changes

-   ba90230: add styling api to embedded checkout
-   Updated dependencies [ba90230]
    -   @whop/checkout@0.0.31

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>