Version Packages (#239)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.32

### Patch Changes

-   e99ce4c: fix malformed style attribute

## @whop/react@0.2.18

### Patch Changes

-   Updated dependencies [e99ce4c]
    -   @whop/checkout@0.0.32

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>