Version Packages (#241)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/framer@0.0.1

### Patch Changes

-   c256a64: extract framer checkout embed into its own package
-   Updated dependencies [c256a64]
    -   @whop/react@0.2.19

## @whop/react@0.2.19

### Patch Changes

-   c256a64: extract framer checkout embed into its own package

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>