Version Packages (#246)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.33

### Patch Changes

-   eda990d: add option to prefill email in checkout embed
-   cf63220: fix onComplete callback

## @whop/framer@0.0.3

### Patch Changes

-   eda990d: add option to prefill email in checkout embed
-   Updated dependencies [eda990d]
-   Updated dependencies [f6b475f]
    -   @whop/react@0.2.20

## @whop/react@0.2.20

### Patch Changes

-   eda990d: add option to prefill email in checkout embed
-   f6b475f: Update frosted-ui to v77
-   Updated dependencies [eda990d]
-   Updated dependencies [cf63220]
    -   @whop/checkout@0.0.33

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>