Version Packages (#251)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.34

### Patch Changes

-   9a6c80c: add option to change checkout embed accent color

## @whop/framer@0.0.4

### Patch Changes

-   9a6c80c: add option to change checkout embed accent color
-   Updated dependencies [9a6c80c]
    -   @whop/react@0.2.21

## @whop/react@0.2.21

### Patch Changes

-   9a6c80c: add option to change checkout embed accent color
-   Updated dependencies [9a6c80c]
    -   @whop/checkout@0.0.34

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>