Version Packages (#253)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.32

### Patch Changes

-   b881027: Add getCourse query

## @whop/framer@0.0.5

### Patch Changes

-   @whop/react@0.2.22

## @whop/react@0.2.22

### Patch Changes

-   Updated dependencies [b881027]
    -   @whop/api@0.0.32

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>