Version Packages (#255)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.33

### Patch Changes

-   44855e8: fix oauth url missing state

## @whop/framer@0.0.6

### Patch Changes

-   @whop/react@0.2.23

## @whop/react@0.2.23

### Patch Changes

-   Updated dependencies [44855e8]
    -   @whop/api@0.0.33

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>