Version Packages (#257)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.34

### Patch Changes

- e9486ac: Add additional fields to courses.getCourse. Create
courses.getLesson, courses.getUserLessonInteractions.

## @whop/framer@0.0.7

### Patch Changes

-   @whop/react@0.2.24

## @whop/react@0.2.24

### Patch Changes

-   Updated dependencies [e9486ac]
    -   @whop/api@0.0.34

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>