Version Packages (#259)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.35

### Patch Changes

- b7181d2: Create courses.markLessonAsCompleted mutation. Add content to
BasicCourseLesson fragment. Add lesson ids to
courses.getUserLessonInteractions.

## @whop/framer@0.0.8

### Patch Changes

-   @whop/react@0.2.25

## @whop/react@0.2.25

### Patch Changes

-   Updated dependencies [b7181d2]
    -   @whop/api@0.0.35

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>