Version Packages (#261)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.36

### Patch Changes

- 4bd6e46: Created initial react native package with build and ship cli

## @whop/framer@0.0.9

### Patch Changes

-   @whop/react@0.2.26

## @whop/react@0.2.26

### Patch Changes

-   Updated dependencies [4bd6e46]
    -   @whop/api@0.0.36

## @whop/react-native@0.0.2

### Patch Changes

- 4bd6e46: Created initial react native package with build and ship cli
-   Updated dependencies [4bd6e46]
    -   @whop/api@0.0.36

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>