Version Packages (#265)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.37

### Patch Changes

- b531da4: Add courses mutations for creating, updating and deleting
lessons and chapters

## @whop/framer@0.0.10

### Patch Changes

-   @whop/react@0.2.27

## @whop/react@0.2.27

### Patch Changes

-   Updated dependencies [b531da4]
    -   @whop/api@0.0.37

## @whop/react-native@0.0.3

### Patch Changes

-   9b240be: Allow interacting with the internal native whop core api
-   Updated dependencies [b531da4]
    -   @whop/api@0.0.37

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>