Version Packages (#271)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.38

### Patch Changes

-   4b77b81: add courses.listCoursesForExperience query

## @whop/framer@0.0.11

### Patch Changes

-   @whop/react@0.2.28

## @whop/react@0.2.28

### Patch Changes

-   Updated dependencies [4b77b81]
    -   @whop/api@0.0.38

## @whop/react-native@0.0.5

### Patch Changes

-   7df4c61: add ability to present sheets and set navigation bar data
-   Updated dependencies [4b77b81]
    -   @whop/api@0.0.38

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>