Add courses.moveCourse mutation. Add courses.listCoursesForCompany query (#275)
Since modules (aka. multiple courses in one experience) were released,
we are adding a section to the courses app that allows creators to move
a course from one experience to another.

For example, if you have an experience with 2 courses, and then have
another experience with 1 course, you could move that course such that
you end up with 3 courses in a single experience.

Adds `listCoursesForCompany` query. We will use this to show the creator
what courses could be moved.

Adds `moveCourse` mutation. We will call this when a creator wants to
move a course.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>