Version Packages (#273)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.39

### Patch Changes

- 31a2054: Change courses.getUserLessonInteractions query to use course
id instead of experience id
- 5a44378: Add courses.moveCourse mutation. Add
courses.listCoursesForCompany query

## @whop/framer@0.0.12

### Patch Changes

-   @whop/react@0.2.29

## @whop/react@0.2.29

### Patch Changes

-   Updated dependencies [31a2054]
-   Updated dependencies [5a44378]
    -   @whop/api@0.0.39

## @whop/react-native@0.0.6

### Patch Changes

-   Updated dependencies [31a2054]
-   Updated dependencies [5a44378]
    -   @whop/api@0.0.39

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>