Version Packages (#277)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.40

### Patch Changes

- c99fb53: switch all course queries to use courseId instead of
experienceId

## @whop/framer@0.0.13

### Patch Changes

-   @whop/react@0.2.30

## @whop/react@0.2.30

### Patch Changes

-   Updated dependencies [c99fb53]
    -   @whop/api@0.0.40

## @whop/react-native@0.0.7

### Patch Changes

-   Updated dependencies [c99fb53]
    -   @whop/api@0.0.40

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>