Version Packages (#288)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.35

### Patch Changes

-   21cfd77: allow embedded checkout to be submitted programmatically

## @whop/framer@0.0.15

### Patch Changes

-   Updated dependencies [21cfd77]
    -   @whop/react@0.2.32

## @whop/react@0.2.32

### Patch Changes

-   21cfd77: allow embedded checkout to be submitted programmatically
-   Updated dependencies [21cfd77]
    -   @whop/checkout@0.0.35

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>