Version Packages (#290)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.42

### Patch Changes

-   be7d17e: Added list receipts for company graphql query
-   d7ebe1b: add auth.js provider to oauth helper

## @whop/framer@0.0.16

### Patch Changes

-   @whop/react@0.2.33

## @whop/react@0.2.33

### Patch Changes

-   Updated dependencies [be7d17e]
-   Updated dependencies [d7ebe1b]
    -   @whop/api@0.0.42

## @whop/react-native@0.0.10

### Patch Changes

-   Updated dependencies [be7d17e]
-   Updated dependencies [d7ebe1b]
    -   @whop/api@0.0.42

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>