Version Packages (#293)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.36

### Patch Changes

-   cd036fe: feat: hide submit button in embedded checkout

## @whop/framer@0.0.17

### Patch Changes

-   Updated dependencies [cd036fe]
    -   @whop/react@0.2.34

## @whop/react@0.2.34

### Patch Changes

-   cd036fe: feat: hide submit button in embedded checkout
-   Updated dependencies [cd036fe]
    -   @whop/checkout@0.0.36

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>