Version Packages (#299)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.37

### Patch Changes

-   6677aae: allow getting state of email field in embedded checkout

## @whop/framer@0.0.18

### Patch Changes

-   Updated dependencies [6677aae]
    -   @whop/react@0.2.35

## @whop/react@0.2.35

### Patch Changes

-   6677aae: allow getting state of email field in embedded checkout
-   Updated dependencies [6677aae]
    -   @whop/checkout@0.0.37

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>