Version Packages (#309)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.38

### Patch Changes

-   d0b1f24: add @whop/checkout/react

## @whop/framer@0.0.19

### Patch Changes

-   @whop/react@0.2.36

## @whop/react@0.2.36

### Patch Changes

-   Updated dependencies [d0b1f24]
    -   @whop/checkout@0.0.38

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>