Version Packages (#318)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/react-native@0.0.13

### Patch Changes

-   ade91e6: added app view type support

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>