Version Packages (#329)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.43

### Patch Changes

-   967283d: Add companies.getCompany graphql query
- db054d6: chore(deps): update dependency @protobuf-ts/protoc to v2.11.1

## @whop/react@0.2.38

### Patch Changes

-   Updated dependencies [967283d]
-   Updated dependencies [db054d6]
    -   @whop/api@0.0.43

## @whop/react-native@0.0.14

### Patch Changes

-   Updated dependencies [967283d]
-   Updated dependencies [db054d6]
    -   @whop/api@0.0.43

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>