Version Packages (#335)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.44

### Patch Changes

-   72f6565: Add bannerImage to company

## @whop/react@0.2.39

### Patch Changes

-   Updated dependencies [72f6565]
    -   @whop/api@0.0.44

## @whop/react-native@0.0.15

### Patch Changes

-   Updated dependencies [72f6565]
    -   @whop/api@0.0.44

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>