Version Packages (#362)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.48

### Patch Changes

- a6079b9: Add getAffiliate query, listAffiliates query, createAffiliate
mutation and updateAffiliate mutation

## @whop/react@0.2.43

### Patch Changes

-   Updated dependencies [a6079b9]
    -   @whop/api@0.0.48

## @whop/react-native@0.0.19

### Patch Changes

-   Updated dependencies [a6079b9]
    -   @whop/api@0.0.48

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>