Version Packages (#365)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/api@0.0.49

### Patch Changes

-   aefdb83: Add create/update/delete access pass mutations
- b07878f: fix: add the optional included error into all responses for
debugging help
- db87a21: Add create/update/get/list app and promote/unassign app build

## @whop/react@0.2.44

### Patch Changes

-   Updated dependencies [aefdb83]
-   Updated dependencies [b07878f]
-   Updated dependencies [db87a21]
    -   @whop/api@0.0.49

## @whop/react-native@0.0.20

### Patch Changes

-   Updated dependencies [aefdb83]
-   Updated dependencies [b07878f]
-   Updated dependencies [db87a21]
    -   @whop/api@0.0.49

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>