crn: add react-native-reanimated (#371)
This PR adds `react-native-reanimated@3.18.0` as a dependency to the
create-react-native template.

Currently running build on a freshly bootstrapped crn app throws this
error:
✔︎ cleaned build directory
✔︎ [ios] entrypoint created
Error: Cannot find module 'react-native-reanimated/plugin'

The plugin is defined in the babel config:

https://github.com/whopio/whop-sdk-ts/blob/40493ce85c48907938bd0e5599957cf58417ca16/packages/react-native/src/cli/babel.config.js#L5

Since we're already in a way "forcing" the dependency, I think we should
just include it for now for a better experience. Ideally we can extend
babel configs.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>