Version Packages (#407)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.41

### Patch Changes

-   d9d1b87: feat: add highContrast to themeOptions

## @whop/framer@0.0.22

### Patch Changes

-   Updated dependencies [d9d1b87]
    -   @whop/checkout@0.0.41

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>