Version Packages (#454)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/react@0.3.0

### Minor Changes

-   0ba95cf: Updated to use expose typescript SDK

## @whop/react-native@0.1.0

### Minor Changes

-   0ba95cf: Updated to use expose typescript SDK

## @whop/checkout@0.0.42

### Patch Changes

-   ce909da: add embedded apple pay support

## @whop/framer@0.0.23

### Patch Changes

-   Updated dependencies [ce909da]
    -   @whop/checkout@0.0.42

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>