feat(checkout): embedded apple pay w/ custom submit (#475)
Apple Pay was not working in embedded checkout when using a custom
submit button. The reason was that the submit button uses postMessage to
submit the embed which does not (reliably) transfer the user gesture
that is needed for the apple pay sheet to show.

This PR introduces handling for that by constructing and showing the
payment request in the host context and then submitting the result with
the submit event.