Version Packages (#476)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @whop/checkout@0.0.43

### Patch Changes

- 2740e4b: feat: support programmatic checkout submission with apple pay

## @whop/framer@0.0.24

### Patch Changes

-   Updated dependencies [2740e4b]
    -   @whop/checkout@0.0.43

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>